+++ CHANGELOG.md
@@ -10,16 +10,61 @@ Versioning].
 
 ## [Unreleased]
 
+## [0.1.3] - 2026-01-17
+
+### kconfq (CLI)
+
+#### Added
+
+- Sub-command `find` to print config's line by its entry name.
+- Sub-command `find-value` to print value of the config's entry.
+
+### kconfq (library)
+
+#### Added
+
+- Function `find_line` to find config's line by its entry name.
+- Function `find_value` to find value of the config's entry.
+- Error `FindLineError`.
+- Error `FindValueError`.
+
+#### Changed
+
+- Rename `GetLinuxKernelVersionError` to `GetKernelVersionError`.
+- Rename `GetLinuxKernelVersionError::MissingUnameRelease` to
+  `GetKernelVersionError::ReleaseMissingFromUname`.
+- Rename `LocateConfigFileError` to `LocateConfigError`.
+- Rename `LocateConfigFileError::ErrorGettingLinuxKernelVersion` to
+  `LocateConfigError::FailedToGetLinuxKernelVersion`.
+- Rename `RequireConfigFileError` to `RequireConfigError`.
+- Rename `RequireConfigFileError::Locate` to
+  `RequireConfigError::FailedToLocate`.
+- Rename `IsGzipError::FailedToReadMagic` to
+  `IsGzipError::FailedToReadFileMagic`.
+
+### libkconfq (C-API)
+
+#### Added
+
+- Opaque struct `KconfConfig`.
+- Function `kconfq_config_path` to get a path to the config's underlying file.
+- Function `kconfq_find_line` to find config's line by its entry name.
+- Function `kconfq_find_value` to find value of the config's entry.
+
+#### Changed
+
+- `kconfq_locate_config` returns `KconfqConfig` instead of path to config file.
+
 ## [0.1.2] - 2026-01-14
 
-### kconfq (Rust CLI)
+### kconfq (CLI)
 
 #### Added
 
 - Sub-command `path` to print path to kernel config.
 - Sub-command `config` to print the whole kernel config.
 
-### kconfq (Rust library)
+### kconfq (library)
 
 #### Added
 
@@ -33,4 +78,5 @@ Versioning].
 - Function `kconfq_locate_config` to find the location of the kernel config.
 
 [Unreleased]: https://github.com/synalice/kconfq/compare/v0.1.2...HEAD
+[0.1.3]: https://github.com/synalice/kconfq/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/synalice/kconfq/releases/tag/v0.1.2